SpecialItem added
There is now a new type of item to interact with.  The SpecialItem is held in the off hand and it can do one of three things: add to attack bonus, add to ac, or allow special actions.  I also fixed a bug in the manage inventory section of Map.py.  The lists of targetable items did not update as objects moved from one place to another, this has been fixed.