Preparing for adding more rooms
I changed the structure of rooms.  Now Room is its own class and GreatHall is-a Room.  Room has the structure that each room will need.  First there is an encounter, then there is a chance to loot the room.  All of the attributes that make the Great Hall unique have been removed from the Room class.