Potions are now useable
I have made changes to Lib.py and Map.py to make potions useable.  Potions are objects that can be equipped in the player's belt.  There are several places where possible locations of objects are tracked and there are several commands that move objects from one place to another.  All of these commands needed to be edited.