Multiple Rooms
There are now two rooms.  The opening scene is in the woods.  There are several items included in the loot from this room.  Once there has been an opportunity to equip and store all of the associated items, there is now an option to enter the Great Hall.  The great hall is the same as before, but with different loot.