Putting this on Hiatus
This game is nowhere close to being fun to play and it has very few items and rooms.  That being said, I feel that I have learned what I need to learn from making it.  It is well past time that I move on with my life.  For one thing, I want to go through the remaining lessons in Learning Python the Hard Way.